animations now are in a separated folder depending the component attached to
